requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Pose Estimation & Computer Vision
# Pin mediapipe to 0.10.14 — last stable version with mp.solutions.holistic API
mediapipe==0.10.14
opencv-python>=4.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Web Application
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0

# Optional: Gemini API for sentence formation
google-generativeai>=0.3.0
